https://lhphong.atlassian.net/browse/CNPM24-64 add createRoom logic; rename file
