+++ src/pages/HomePage/HomePage.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@mui/material";
 import React, { useEffect } from "react";
 import { useAuth } from "../../lib/provider/AuthProvider";
-import AddRoomButton from "../../modules/room/components/Dialog/AddRoomButton";
+import AddRoomButton from "../../modules/room/components/AddRoomButton/AddRoomButton";
 import RoomItem from "../../modules/room/components/RoomItem/RoomItem";
 
 const HomePage = () => {
